chore(usedoc): move usedoc
