+++ app/src/components/GraphContextMenu.tsx
@@ -19,8 +19,8 @@ import { useThemeStore } from "../lib/graphThemes";
 import { borderStyles, shapes } from "../lib/graphUtilityClasses";
 import { useIsFirefox } from "../lib/hooks";
 import { useParser } from "../lib/parsers";
-import { useDoc } from "../lib/prepareChart";
 import { useContextMenuState } from "../lib/useContextMenuState";
+import { useDoc } from "../lib/useDoc";
 import { Box, Type } from "../slang";
 import styles from "./GraphContextMenu.module.css";
 import { smallIconSize } from "./Shared";
