+++ app/src/pages/Edit.tsx
@@ -18,8 +18,8 @@ import { EditStyleTab } from "../components/Tabs/EditStyleTab";
 import { TextEditor } from "../components/TextEditor";
 import { titleToLocalStorageKey } from "../lib/helpers";
 import { useIsValidCustomer, useIsValidSponsor } from "../lib/hooks";
-import { Doc, docToString, prepareChart, useDoc } from "../lib/prepareChart";
 import { getDefaultText } from "../lib/queries";
+import { Doc, docToString, prepareChart, useDoc } from "../lib/useDoc";
 import { useTrackLastChart } from "../lib/useLastChart";
 import { Type } from "../slang";
 import styles from "./Edit.module.css";
