+++ app/src/pages/EditHosted.tsx
@@ -19,8 +19,8 @@ import { EditMetaTab } from "../components/Tabs/EditMetaTab";
 import { EditStyleTab } from "../components/Tabs/EditStyleTab";
 import { TextEditor } from "../components/TextEditor";
 import { useIsValidSponsor } from "../lib/hooks";
-import { Doc, docToString, prepareChart, useDoc } from "../lib/prepareChart";
 import { getHostedChart, updateChartText } from "../lib/queries";
+import { Doc, docToString, prepareChart, useDoc } from "../lib/useDoc";
 import { useTrackLastChart } from "../lib/useLastChart";
 import editStyles from "./Edit.module.css";
 import styles from "./EditHosted.module.css";
