+++ app/src/pages/Public.tsx
@@ -5,8 +5,8 @@ import { CloneButton } from "../components/CloneButton";
 import Graph from "../components/Graph";
 import GraphWrapper from "../components/GraphWrapper";
 import styles from "../components/Main.module.css";
-import { prepareChart } from "../lib/prepareChart";
 import { getPublicChart } from "../lib/queries";
+import { prepareChart } from "../lib/useDoc";
 
 const shouldResize = 0;
