+++ app/src/pages/ReadOnly.tsx
@@ -9,7 +9,7 @@ import { EditorWrapper } from "../components/EditorWrapper";
 import { EditWrapper } from "../components/EditWrapper";
 import Main from "../components/Main";
 import { TextEditor } from "../components/TextEditor";
-import { prepareChart, useDoc } from "../lib/prepareChart";
+import { prepareChart, useDoc } from "../lib/useDoc";
 
 function ReadOnly() {
   const { path } = useRouteMatch();
